refactor(lotto): ChangeInput와 InputView 병합
